Adding router with blueprint
