+++ src/routes.py
@@ -1,4 +1,14 @@
-from flask import Blueprint, render_template, session, current_app
+from flask import (
+    Blueprint,
+    current_app,
+    redirect,
+    render_template,
+    request,
+    session,
+    url_for,
+)
+
+from src.auth import authorize_url
 
 router = Blueprint("router", __name__)
 
@@ -15,3 +25,15 @@ def app_main():
         return render_template("index.html")
 
     return render_template("index.html", athlete_id=obj.session["athlete_id"])
+
+
+# Route for handling the login page logic
+@router.route("/login", methods=["GET", "POST"])
+def login():
+    error = None
+    if request.method == "POST":
+        if request.form["username"] != "admin" or request.form["password"] != "admin":
+            error = "Invalid Credentials. Please try again."
+        else:
+            return redirect(url_for("home"))
+    return render_template("base.html", error=error)
